cqured DOC
Both cqured and coreFramework's DOC are very similar in terms of structure .
They are expected to share a single DOC for models and CoreModel query .
airMedia Doc almost completed. Left with adding the audio doc.

We expect to give background images for all the banners of the individual docs.

Home page is expected to be polished.

Create a component to show show follow me on github.